fix(h7): address CodeRabbitAI review findings from PR #1194
Resolves all actionable items from the large inline + major/minor review
on the stm32h7-support backport PR. All fixes maintain parity with
BF 4.5-maintenance while adapting to EmuFlight architecture.

Drivers:
- adc.c: DCache invalidation size hardcoded 32 → sizeof(adcValues) rounded
- adc_stm32h7xx.c: use RCC_ClockCmd(adc.rccADC) instead of hardcoded ADC12; add STM32H735xx to VREFINT 12-bit shift group
- bus_i2c_timing.c: I2C prescaler loop off-by-one (< 15 → <= 15)
- bus_octospi_stm32h7xx.c: MODIFY_REG CCR config written to ABR → CCR
- bus_quadspi.h: QUADSPI_CLOCK_INITIALIZATION 256 → 255
- bus_spi_ll.c: dummyTxByte/dummyRxByte → DMA_DATA/DMA_DATA_ZERO_INIT; add txData DTCM safety check
- dma_reqmap_mcu.c: add STM32H735xx to ADC3 guard; replace dmaoptByTag (always DMA_OPT_UNUSED on H7) with dmaGetOptionByTimer; reorder to eliminate forward reference
- dma_stm32h7xx.c: add __HAL_RCC_DMAMUX1_CLK_ENABLE() in enableDmaClock()
- exti.c: EXTI active mask missing & 0xFFFFu (H7 IMR1/PR1 upper bits are reserved)
- sdio.h: add #pragma once
- sdio_h7xx.c: move TXCplt=1 after validation; clear TXCplt/RXCplt on HAL error
- serial_uart.h: add USE_UART9/UART10 to USE_UART gate
- system.c: pointer subtraction between linker symbols → uintptr_t cast
- system_stm32h7xx.c: enable GPIOA/GPIOC clocks before GPIO init in configureMasterClockOutputs()
- timer_def.h: add DEF_TIM_DMA_REQ__BTCH_TIM12/13/14_* macros (DMA_REQUEST_NONE, no H7 DMAMUX slots)
- transponder_ir_io_hal.c: H7 GPIO speed guard in transponderIrDisable()

Startup / linker:
- startup_stm32h730xx.s: add bl persistentObjectInit before SystemInit (matches H735)
- stm32_flash_h7a3_2m.ld: fix .data/.bss VMA FASTRAM → RAM
- stm32_ram_h723_exst.ld: _estack missing -8 reserved slot
- stm32_ram_h723_exst_post.ld: typo "MD5 has is" → "MD5 hash is"
- stm32_ram_h730_exst.ld: _estack missing -8 reserved slot
- stm32_ram_h735_exst.ld: INCLUDE stm32_h723_common.ld → stm32_h735_common.ld
- stm32_ram_h750_exst.ld: _estack missing -8 reserved slot

USB:
- usbd_conf_stm32h7xx.c: add OTG_HS_EP1_OUT/IN IRQ handlers for H730; disable EP1 IRQs in MspDeInit

+++ src/main/drivers/system.c
@@ -196,13 +196,13 @@ void initialiseMemorySections(void)
     extern uint8_t tcm_code_start;
     extern uint8_t tcm_code_end;
     extern uint8_t tcm_code;
-    memcpy(&tcm_code_start, &tcm_code, (size_t) (&tcm_code_end - &tcm_code_start));
+    memcpy(&tcm_code_start, &tcm_code, (size_t) ((uintptr_t)&tcm_code_end - (uintptr_t)&tcm_code_start));
 #endif
 #ifdef USE_FAST_DATA
     extern uint8_t _sfastram_data;
     extern uint8_t _efastram_data;
     extern uint8_t _sfastram_idata;
-    memcpy(&_sfastram_data, &_sfastram_idata, (size_t) (&_efastram_data - &_sfastram_data));
+    memcpy(&_sfastram_data, &_sfastram_idata, (size_t) ((uintptr_t)&_efastram_data - (uintptr_t)&_sfastram_data));
 #endif
 }
 
@@ -214,7 +214,7 @@ void initialiseD2MemorySections(void)
     extern uint8_t _sdmaram_data;
     extern uint8_t _edmaram_data;
     extern uint8_t _sdmaram_idata;
-    memset(&_sdmaram_bss, 0, (size_t) (&_edmaram_bss - &_sdmaram_bss));
-    memcpy(&_sdmaram_data, &_sdmaram_idata, (size_t) (&_edmaram_data - &_sdmaram_data));
+    memset(&_sdmaram_bss, 0, (size_t) ((uintptr_t)&_edmaram_bss - (uintptr_t)&_sdmaram_bss));
+    memcpy(&_sdmaram_data, &_sdmaram_idata, (size_t) ((uintptr_t)&_edmaram_data - (uintptr_t)&_sdmaram_data));
 }
 #endif
